+++ docs/CHANGE_LOG.md
@@ -15,7 +15,7 @@
 #### **Public Functions:**
 - Added a new public function "moveToPreviousCategory()", which will move to the next category for a specific element and then refresh its UI.
 - Added a new public function "moveToNextCategory()", which will move to the previous category for a specific element and then refresh its UI.
-- Added a new public function "refresh()", which which refresh the UI for a specific element.
+- Added a new public function "refresh()", which refreshes the UI for a specific element.
 - Added a new public function "refreshAll()", which will refresh all the rendered elements.
 
 #### **Configuration Options:**